+++ src/Set/Properties.php
@@ -53,7 +53,19 @@ final class Properties implements Provider
      */
     public function atMost(int $max): Set
     {
-        return $this->ensure($max);
+        return $this->ensure(null, $max);
+    }
+
+    /**
+     * @psalm-mutation-free
+     *
+     * @param positive-int $min
+     *
+     * @return Set<Ensure>
+     */
+    public function atLeast(int $min): Set
+    {
+        return $this->ensure($min);
     }
 
     /**
@@ -116,20 +128,24 @@ final class Properties implements Provider
     #[\Override]
     public function toSet(): Set
     {
-        return $this->ensure(100);
+        return $this->ensure();
     }
 
     /**
      * @psalm-mutation-free
      *
+     * @param positive-int $min
      * @param positive-int $max
      *
      * @return Set<Ensure>
      */
-    private function ensure(int $max): Set
+    private function ensure(?int $min = null, ?int $max = null): Set
     {
+        $min ??= 1;
+        $max ??= $min + 99;
+
         /** @var Set<non-empty-list<Concrete>> */
-        $sequences = Set::sequence($this->properties)->between(1, $max);
+        $sequences = Set::sequence($this->properties)->between($min, $max);
 
         return $sequences->map(
             static fn(array $properties): Ensure => Ensure::of(...$properties),